# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Adaption Tracker",
        "contact": {
            "email": "m.mohamed@cat.com.eg"
        },
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "login",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "a@a.com"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "logout currently logged in user",
                "responses": {
                    "200": {
                        "description": "Successfully logged out"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/refresh": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "refreshes expired token",
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/forms": {
            "get": {
                "tags": [
                    "forms"
                ],
                "summary": "Get forms",
                "responses": {
                    "200": {
                        "description": "objects"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/forms/{form}": {
            "get": {
                "tags": [
                    "forms"
                ],
                "summary": "Get single form",
                "parameters": [
                    {
                        "name": "form",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/forms/{form}/submit": {
            "post": {
                "tags": [
                    "forms"
                ],
                "summary": "submit form",
                "parameters": [
                    {
                        "name": "form",
                        "in": "path",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/me": {
            "get": {
                "tags": [
                    "account"
                ],
                "summary": "Get the current logged in user",
                "responses": {
                    "200": {
                        "description": "object"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Account",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "mahmoud"
                    },
                    {
                        "name": "phone",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "01205528825"
                    },
                    {
                        "name": "enable_questions",
                        "in": "formData",
                        "required": true,
                        "type": "integer",
                        "default": "0"
                    },
                    {
                        "name": "bio",
                        "in": "formData",
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string",
                        "default": "m.mohamed@cat.com.eg"
                    },
                    {
                        "name": "photo",
                        "in": "formData",
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/account/update-password": {
            "post": {
                "tags": [
                    "account"
                ],
                "summary": "update My Password",
                "parameters": [
                    {
                        "name": "type",
                        "in": "header",
                        "description": "1 => attendee , 2=> speaker",
                        "required": true,
                        "type": "integer",
                        "format": "integer"
                    },
                    {
                        "name": "current_password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "required": true,
                        "type": "string",
                        "format": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Model"
                    },
                    "400": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Value: Bearer \\<token\\><br><br>",
            "name": "Authorization",
            "in": "header"
        }
    }
}